Fix desktop file rewriting
When switching from `build-commands` to `post-install` during the review process the desktop file exec rewriting process began editing the source repo desktop file after it was already installed. This would mean that the actual desktop file in the Flatpak would not get rewritten. This change edits the installed desktop file instead of the source file.